requirements.txt
PyMuPDF>=1.24.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
streamlit>=1.38.0
reportlab>=4.0.0
pysqlite3-binary
python-docx
